feat: Support Native Animation LTPO feature



# message auto-generated for no-merge-commit merge:
merge newltpo into 0.77.1-rc.1-ohos

feat: Support Native Animation LTPO feature

Created-by: panjinling
Commit-by: panjinling
Merged-by: huangyouhua
Description: ## Description
The OH_DisplaySoloist_* series of interfaces are used to control the display refresh rate (through a voting mechanism). When React Native (RN) runs native animations, it calculates the speed/transformation rate of the animations and then converts them into the corresponding display refresh rate through a formula, thereby controlling the display refresh rate.


## Changes
feat: Support Native Animation LTPO feature

## Test Plan

The following three code snippets are used to test the interface frame rate when the frameAnimation, springAnimation, and decayAnimation run at different speeds. When the running speed changes, the frame rate (which can be viewed by enabling the developer mode) changes accordingly.

case1:
const startAnimation = () => {
    Animated.spring(translateX, {
      toValue: 200,
      friction: 5,         // 弹簧摩擦系数（越大阻力越大）
      tension: 40,         // 弹性张力（越大反弹越快）
      useNativeDriver: true, // ✅ 使用原生驱动
    }).start();
  };


case2:
    useEffect(() => {
    Animated.loop(
      Animated.sequence([
        Animated.spring(translateX, {
          toValue: 200,
          friction: 5,         // 弹簧摩擦系数（越大阻力越大）
          tension: 40,         // 弹性张力（越大反弹越快）
          useNativeDriver: true, // ✅ 使用原生驱动
        }),
        Animated.spring(translateX, {
          toValue: -200,
          friction: 5,         // 弹簧摩擦系数（越大阻力越大）
          tension: 40,         // 弹性张力（越大反弹越快）
          useNativeDriver: true, // ✅ 使用原生驱动
        }),
      ])
    ).start();
  }, []);
  
  
case3:
 useEffect(() => {
    Animated.loop(
      Animated.sequence([
        Animated.decay(translateX, {
          velocity: 1,              // 初速度（正数向右，负数向左）
          deceleration: 0.992,      // 衰减因子（接近 1 时滑得更远）
          useNativeDriver: true,    // ✅ 使用原生驱动
        }),
        Animated.decay(translateX, {
          velocity: -1,              // 初速度（正数向右，负数向左）
          deceleration: 0.992,      // 衰减因子（接近 1 时滑得更远）
          useNativeDriver: true,    // ✅ 使用原生驱动
        }),
      ])
    ).start(({finished, value})=>{
      console.log("xxx value = " + value)
    });
  }, []);

  const startDecay = () => {
    Animated.decay(translateX, {
      velocity: 1,              // 初速度（正数向右，负数向左）
      deceleration: 0.992,      // 衰减因子（接近 1 时滑得更远）
      useNativeDriver: true,    // ✅ 使用原生驱动
    }).start();
  };



## Checklist
- [x] Does not involve incompatible changes; if involved, has been reviewed accordingly.
- [x] Does not impact performance, or performance testing has been conducted without degradation.
- [x] Complies with the relevant coding standards.
- [x] Does not involve documentation updates, or the documentation has been updated.
- [x] Meets testability requirements with necessary self-test cases, appropriate logging, or trace information added.
- [x] No illegal file inclusions exist, such as images or code.


See merge request: openharmony-sig/ohos_react_native!1195

+++ packages/tester/harmony/react_native_openharmony/src/main/cpp/RNOHCorePackage/TurboModules/Animated/Drivers/AnimationDriver.h
@@ -30,7 +30,7 @@ class AnimationDriver {
     }
   };
 
-  virtual void runAnimationStep(uint64_t frameTimeNanos) = 0;
+  virtual void runAnimationStep(long long frameTimeNanos) = 0;
   virtual void resetConfig(folly::dynamic const& config);
 
   facebook::react::Tag getId() const;
